Mkdir module-1

+++ module-1/p7.cpp
@@ -3,11 +3,11 @@
 using namespace std;
 
 int main () {
-    double temp;
+    float temp;
     cout << "What is your temperature in Fahrenheit?: ";
     cin >> temp;
 
-    temp = 5 / 9 * (temp-32);
+    temp = 5.0 / 9 * (temp-32);
 
     cout << "\nThe temperature in Celsius is: " << temp << endl;
 }
